Basic changes
Since we want to use Sprites, which we manipulate through our classes and a game controller I have adjusted where some variables are.
This for now is that the player no longer has a constructor, and the position (and vector) values are used from the sprite class.
The changes on GameScreen and LevelData are experimental, and don't do anything yet.