Added some files, and changed some classes
Added a json file for a test map and a Tileset.
Added the pragma once thing.
Made Memories, Dirt and Boulder inherit from Sprite.
Sprite now has a rectangle for the hurtbox, not sure if this will be useful.
Adjusted the Screensize, and scaling of textures (instead of pointlessly enhancing everything twice, i reduced the screensize)
--> The Screen has some weird problems when going into fullscreen.
Am trying to get the Gamescreen class to work, aswell as the LevelData, This might take a while.